convert class component to functional component
